% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_agendas_and_minutes.R
\name{download_data_from_meeting_page}
\alias{download_data_from_meeting_page}
\title{Download agenda and/or minutes files from a meeting link page scraped from Framingham meeting portal}
\usage{
download_data_from_meeting_page(
  the_meeting_link,
  the_root = "http://framinghamma.iqm2.com",
  the_dir = "tmp"
)
}
\arguments{
\item{the_meeting_link}{character string}

\item{the_root}{character string with website root}

\item{the_dir}{character string with directory where you want files downloaded}
}
\value{
data frame with info about meeting files downloaded
}
\description{
Download agenda and/or minutes files from a meeting link page scraped from Framingham meeting portal
}
