% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_agendas_and_minutes.R
\name{get_list_of_meeting_links}
\alias{get_list_of_meeting_links}
\title{Get list of meeting page links from Framingham meeting portal}
\usage{
get_list_of_meeting_links(
  the_url = "http://framinghamma.iqm2.com",
  the_start = "1/1/2021",
  the_end = "12/31/2021"
)
}
\arguments{
\item{the_url}{string root of URL for meeting listing}

\item{the_start}{string date in m/d/yyyy format}

\item{the_end}{string date in m/d/yyyy format}
}
\value{
vector of character strings with URLs of meeting pages
}
\description{
Get list of meeting page links from Framingham meeting portal
}
