% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_pages}
\alias{get_pages}
\title{Get number of pages in a PDF file}
\usage{
get_pages(the_file)
}
\arguments{
\item{the_file}{character string with name of file}
}
\value{
integer with number of pages in the PDF
}
\description{
Get number of pages in a PDF file
}
