% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_text_from_pdf}
\alias{get_text_from_pdf}
\title{Convert meeting agenda or minutes to text. Limit to max number of pages if desired}
\usage{
get_text_from_pdf(the_file, max_pages = 0)
}
\arguments{
\item{the_file}{character string name of file}

\item{max_pages}{integer maxinum number of pages of text to return. If 0, return all}
}
\value{
character string with text of PDF
}
\description{
Convert meeting agenda or minutes to text. Limit to max number of pages if desired
}
