% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_agendas_and_minutes2.R
\name{get_list_of_meeting_links}
\alias{get_list_of_meeting_links}
\title{Get list of meetings, agendas, minutes, and links from Framingham meeting portal}
\usage{
get_list_of_meeting_links(
  the_url = "http://framinghamma.iqm2.com",
  the_start = "1/1/2021",
  the_end = "12/31/2021",
  my_dir = "www"
)
}
\arguments{
\item{the_url}{string root of URL for meeting listing}

\item{the_start}{string date in m/d/yyyy format}

\item{the_end}{string date in m/d/yyyy format}

\item{my_dir}{string subdirectory you want files downloaded; set to "" for none}
}
\value{
data frame with URLs of meeting pages, agendas, and minutes as well as meeting info and separate columns for Date and Board (City Council, etc.)
}
\description{
Get list of meetings, agendas, minutes, and links from Framingham meeting portal
}
